Added feature and refactored sap service implementations
