+++ src/rest/model/operationType.ts
@@ -11,7 +11,7 @@
  */
 
 
-export type OperationType = 'SAP_CONTACT_SYNC' | 'SAP_DELIVERY_PLACE_SYNC' | 'SAP_ITEM_GROUP_SYNC' | 'SAP_CONTRACT_SYNC' | 'SAP_CONTRACT_SAPID_SYNC' | 'ITEM_GROUP_DEFAULT_DOCUMENT_TEMPLATES';
+export type OperationType = 'SAP_CONTACT_SYNC' | 'SAP_DELIVERY_PLACE_SYNC' | 'SAP_ITEM_GROUP_SYNC' | 'SAP_CONTRACT_SYNC' | 'SAP_CONTRACT_SAPID_SYNC' | 'ITEM_GROUP_DEFAULT_DOCUMENT_TEMPLATES' | 'UPDATE_CURRENT_YEAR_APPROVED_CONTRACTS_TO_SAP';
 
 export const OperationType = {
     SAPCONTACTSYNC: 'SAP_CONTACT_SYNC' as OperationType,
@@ -19,5 +19,6 @@ export const OperationType = {
     SAPITEMGROUPSYNC: 'SAP_ITEM_GROUP_SYNC' as OperationType,
     SAPCONTRACTSYNC: 'SAP_CONTRACT_SYNC' as OperationType,
     SAPCONTRACTSAPIDSYNC: 'SAP_CONTRACT_SAPID_SYNC' as OperationType,
-    ITEMGROUPDEFAULTDOCUMENTTEMPLATES: 'ITEM_GROUP_DEFAULT_DOCUMENT_TEMPLATES' as OperationType
+    ITEMGROUPDEFAULTDOCUMENTTEMPLATES: 'ITEM_GROUP_DEFAULT_DOCUMENT_TEMPLATES' as OperationType,
+    UPDATECURRENTYEARAPPROVEDCONTRACTSTOSAP: 'UPDATE_CURRENT_YEAR_APPROVED_CONTRACTS_TO_SAP' as OperationType
 };
